chore: dto 타입 변경 외 4

+++ src/main/java/market/fundingmarket/domain/sponsorship/dto/response/SponsorResponse.java
@@ -13,7 +13,7 @@ import java.util.UUID;
 @Getter
 public class SponsorResponse {
     private final Long id;
-    private final Long amount;
+    private final int amount;
     private final LocalDateTime sponsoredAt;
     private final int quantity;
     private final boolean cancelled;
